+++ client/src/components/timetable/Timetable.tsx
@@ -7,9 +7,8 @@ import TimetableLayout from './TimetableLayout';
 import ClassDropzones from './ClassDropzones';
 import DroppedClasses from './DroppedClasses';
 import Inventory from '../inventory/Inventory';
-import CourseSelect from '../CourseSelect';
 
-const rowHeight = 92;
+const rowHeight = 85;
 
 const StyledTimetable = styled(Box) <{
   rows: number
@@ -18,20 +17,10 @@ const StyledTimetable = styled(Box) <{
   min-height: ${(props) => props.rows * rowHeight}px;
   max-height: ${(props) => props.rows * rowHeight}px; // TODO: should be different to min-height
   margin-top: 15px;
-  margin-bottom: 15px;
   box-sizing: content-box;
 
   grid-gap: ${1 / devicePixelRatio}px;
-  grid-template: auto repeat(${(props) => 2 * props.rows + 1}, 1fr) auto / auto repeat(${days.length}, 1fr) 11px 1fr;
-`;
-
-const SelectWrapper = styled(Box)`
-  display: flex;
-  flex-direction: row;
-  grid-column: 1 / -1;
-  grid-row: 1;
-  padding-top: 5px;
-  padding-bottom: 10px;
+  grid-template: repeat(${(props) => 2 * props.rows + 1}, 1fr) auto / auto repeat(${days.length}, 1fr) 11px 1fr;
 `;
 
 interface TimetableProps {
@@ -42,8 +31,6 @@ interface TimetableProps {
   setIs12HourMode(value: boolean): void
   onSelectClass(classData: ClassData): void
   onRemoveClass(classData: ClassData): void
-  onSelectCourse(courseCode: string): void
-  onRemoveCourse(courseCode: string): void
 }
 
 const Timetable: FunctionComponent<TimetableProps> = ({
@@ -54,22 +41,12 @@ const Timetable: FunctionComponent<TimetableProps> = ({
   setIs12HourMode,
   onSelectClass,
   onRemoveClass,
-  onSelectCourse,
-  onRemoveCourse,
 }) => (
   <StyledTimetable
     rows={Math.max(...selectedCourses.map(
       (course) => course.latestFinishTime,
     ), defaultEndTime) - defaultStartTime}
   >
-    <SelectWrapper>
-      <CourseSelect
-        selectedCourses={selectedCourses}
-        assignedColors={assignedColors}
-        handleSelect={onSelectCourse}
-        handleRemove={onRemoveCourse}
-      />
-    </SelectWrapper>
     <Inventory
       key={selectedCourses.map((course) => course.code).join(',')}
       selectedCourses={selectedCourses}
